+++ extensions/Exporters/SmugMugExport/SmugMugExport.cs
@@ -550,7 +550,7 @@ namespace FSpotSmugMugExport {
 			filters.Add (new JpegFilter ());
 
 			if (scale)
-				filters.Add (new ResizeFilter ((uint)size));
+				filters.Add (new ResizeFilter (size));
 
 			if (rotate)
 				filters.Add (new OrientationFilter ());
